chore(BV, CP): Refactor propagation mechanism
This patch simplifies the propagation mechanism (currently used in the
bit-vector relations only) in order to accomodate different types of
propagators more easily.

In particular, there is now a single generic (and configurable) loop
that runs the propagators instead of a spaghetti of different loops for
each kind of propagators.