Remove My_zip module (#1264)
* Use similar test for zip feature in native and SMT format

We should use the same test as we only want to check that Alt-Ergo can
unzip correctly the input file if its name has the form:
name.format.zip.

* Remove My_zip

The module `My_Zip` contains only function which is called once in
`Solving_loop`. This commit cleans the code and move it in
`Solving_loop` directly. Technically, `Solving_loop` is a part of the
binary, but as everyone knows, most of the code of this module will be
push into the library after we design a good API.

* Force deflate compression in zip tests

Add a padding comment in both tests to ensure that zip won't choose
the store method while compressing them.